👌 move synthetics context module in the domain folder
