+++ packages/rum-core/src/boot/rumPublicApi.ts
@@ -28,7 +28,7 @@ import { RumSession } from '../domain/rumSession'
 import { RumEventDomainContext } from '../domainContext.types'
 import { CommonContext, User, ActionType, ReplayStats } from '../rawRumEvent.types'
 import { RumEvent } from '../rumEvent.types'
-import { willSyntheticsInjectRum } from '../tools/syntheticsContext'
+import { willSyntheticsInjectRum } from '../domain/syntheticsContext'
 import { buildEnv } from './buildEnv'
 import { startRum } from './startRum'
