+++ packages/rum-core/src/domain/assembly.ts
@@ -24,7 +24,7 @@ import {
   User,
 } from '../rawRumEvent.types'
 import { RumEvent } from '../rumEvent.types'
-import { getSyntheticsContext } from '../tools/syntheticsContext'
+import { getSyntheticsContext } from './syntheticsContext'
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { ParentContexts } from './parentContexts'
 import { RumSession, RumSessionPlan } from './rumSession'
