+++ packages/rum-core/src/domain/syntheticsContext.spec.ts
