+++ packages/rum-core/src/domain/syntheticsContext.ts
